+++ apps/explorer/config/runtime/test.exs
@@ -19,6 +19,8 @@ config :explorer, Explorer.Market.History.Historian, enabled: false
 
 config :explorer, Explorer.Counters.AddressesCounter, enabled: false, enable_consolidation: false
 config :explorer, Explorer.Counters.LastOutputRootSizeCounter, enabled: false, enable_consolidation: false
+config :explorer, Explorer.Counters.Transactions24hStats, enabled: false, enable_consolidation: false
+config :explorer, Explorer.Counters.FreshPendingTransactionsCounter, enabled: false, enable_consolidation: false
 config :explorer, Explorer.Chain.Cache.ContractsCounter, enabled: false, enable_consolidation: false
 config :explorer, Explorer.Chain.Cache.NewContractsCounter, enabled: false, enable_consolidation: false
 config :explorer, Explorer.Chain.Cache.VerifiedContractsCounter, enabled: false, enable_consolidation: false